backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-multipart==0.0.6
aiosmtplib==3.0.1
email-validator==2.1.0
razorpay==1.4.1
